Phase 13: kernel profiling results + HYP-031 + 13a fixes
Three parallel changes:

1. HYP-030 updated with standalone kernel-vs-kernel data:
   TQ beats FlashInfer at seq≤256 (0.34x at seq=128), but scales to
   12.3x slower at seq=4096. FlashInfer is flat (~0.21ms) because
   tensor cores hide O(seq) work. TQ's scalar FMA is the bottleneck.
   Kernel is only 11% of vLLM TPOT — 89% is framework overhead.

2. HYP-031 drafted: tensor-core dequant hypothesis. Marlin-style
   dequant→fp16→mma.sync to flatten TQ's latency curve. Predicted:
   seq=4096 drops from 2.6ms to ~0.3-0.5ms. Status: pending.

3. 13a Python overhead: removed 3 redundant .to(int32) calls from
   the decode forward() hot path. seq_lens, block_table, and
   kv_last_page_len are already int32 from vLLM's allocators.

+++ turboquant/vllm_backend_fused.py
@@ -250,8 +250,8 @@ class TurboQuantFusedImpl(FlashAttentionImpl):
             kv_indptr = (torch.arange(seq_lens.shape[0] + 1,
                                       dtype=torch.int32, device=q.device)
                          * max_pages)
-            kv_last_page_len = (seq_lens - (num_pages - 1) * block_size).to(torch.int32)
-            kv_indices = block_table.reshape(-1).to(torch.int32)
+            kv_last_page_len = seq_lens - (num_pages - 1) * block_size
+            kv_indices = block_table.reshape(-1)
 
             q_fp16 = q.to(torch.float16)
             if self.head_size < self._pd:
@@ -264,7 +264,7 @@ class TurboQuantFusedImpl(FlashAttentionImpl):
             result = torch.ops.turboquant.decode_v4_from_cache(
                 q_fp16, kv_cache,
                 kv_indices, kv_indptr, kv_last_page_len,
-                seq_lens.to(torch.int32),
+                seq_lens,
                 self.num_kv_heads, block_size,
                 self.head_size, self._pd, self.scale,
                 self._signs,
